[Tmdb] Move API files into tmdb/ folder
