+++ src/scrapers/tv_show/tmdb/CMakeLists.txt
@@ -11,7 +11,7 @@ add_library(
 )
 
 target_link_libraries(
-  mediaelch_scraper_tv_tmdb PRIVATE mediaelch_scraper_scraper_apis Qt5::Sql
+  mediaelch_scraper_tv_tmdb PRIVATE mediaelch_scraper_scraper_tmdb Qt5::Sql
                                     Qt5::Widgets Qt5::Multimedia Qt5::Xml
 )
 mediaelch_post_target_defaults(mediaelch_scraper_tv_tmdb)
