+++ src/scrapers/tv_show/tmdb/TmdbTvEpisodeParser.cpp
@@ -1,7 +1,7 @@
 #include "scrapers/tv_show/tmdb/TmdbTvEpisodeParser.h"
 
 #include "globals/Poster.h"
-#include "scrapers/api/TmdbApi.h"
+#include "scrapers/tmdb/TmdbApi.h"
 #include "tv_shows/TvDbId.h"
 #include "tv_shows/TvShowEpisode.h"
